Explicitly call start_process when a restart_command is not defined
Waiting for the the tick to start processes that have been stopped by the restart_process function breaks the ability for one to restart processes in a blocking fashion.

A user who wants to rely on the stop/start that is used when a restart_command is not defined, should get the proper stop/start within the "bluepill restart" command. This method shouldn't rely on the natural bluepill tick to bring processes back up.